Udemy Project Analysis With Presentation
I did an In-depth analysis to better understand the pricing and subscriptions data of Udemy courses (for four different subjects) and identified trends and patterns. 
The objective is to help us:
1. Identify if the courses content duration impact the price of courses , as the price of courses and subscription impacts greatly on company revenue.
2. Identify for which subject more courses should be created.
3. Create targeted strategies to increase the company's revenue.
I used statistical methods, root cause analysis and analytical problem-solving techniques to solve these and proffered recommended action/solution to stakeholders.